Main/DHCPConfigImpl.cpp: Signedness warnings. bugref:9790
svn:sync-xref-src-repo-rev: r139211

+++ src/VBox/Main/src-server/DHCPConfigImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: DHCPConfigImpl.cpp 82968 2020-02-04 10:35:17Z knut.osmundsen@oracle.com $ */
+/* $Id: DHCPConfigImpl.cpp 85234 2020-07-11 16:31:44Z knut.osmundsen@oracle.com $ */
 /** @file
  * VirtualBox Main - IDHCPConfig, IDHCPConfigGlobal, IDHCPConfigGroup, IDHCPConfigIndividual implementation.
  */
@@ -275,9 +275,9 @@ HRESULT DHCPConfig::i_setForcedOptions(const std::vector<DHCPOption_T> &aOptions
         /* Actually changed? */
         if (m_vecForcedOptions.size() == aOptions.size())
         {
-            ssize_t i = m_vecForcedOptions.size();
+            ssize_t i = (ssize_t)m_vecForcedOptions.size();
             while (i-- > 0)
-                if (m_vecForcedOptions[i] != aOptions[i])
+                if (m_vecForcedOptions[(size_t)i] != aOptions[(size_t)i])
                     break;
             if (i < 0)
                 return S_OK;
